Split actions and selectors into unique files
